+++ maptasker/src/primitem.py
@@ -164,7 +164,6 @@ class PrimeItems:
         "Czech": "cs",
     }
     languages_translated: ClassVar[dict[str, str]] = {}
-    no_profile_translated_length: ClassVar[int] = 0
 
 
 # Reset all values