+++ Changelog.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file!
 - Fixed: Task action limit gets reset to 100 in the GUI when changing to another language.
 - Fixed: Clicking on a Task, in the Map view, inadvertantly jumps to the 'Launcher Task' line if there is one rather than to the Task.
 - Fixed: If highlighting Tasker objects (e.g. selecting 'Bold') in trhe Map view, then hotlinks to Tasker objects would result in the error message 'Could not find ...'.
+- Fixed: The Diagram view boxes for 'No Profile' tasks are not the proper length if the language is other than English.
 
 ### Known Issue